RC1: Added pause and some menus
There were problems with static fields not getting reset when the scene is restarted as it turned out.

The menus are on separate canvases and work nicely.

There is a bug where the game would start running (GameState = Running), but some callbacks were not properly attached leading to the canvas not hiding itself and the music not starting.

00:12
13/7/20